Move run_and_event_writer into its own module
